pyproject.toml
[project]
name = "poc-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.4.2",
    "python-dotenv>=1.1.0",
]
